requirements.txt
flask
pandas
openpyxl
xmltodict
python-dotenv
google-generativeai